+++ Commands/ClientSidePages/ClientSidePageTransformatorLogType.cs
@@ -6,6 +6,7 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
         None = 0,
         File = 1,
         SharePoint = 2,
+        Console = 3,
     }
 }
 #endif